feat: Apply link underline on to content links.
This prevents wackiness of underlines spawning into the wp-admin toolbar and nonsense.

We're also forcing buttons to override this behaviour by styling the <a> element directly. Usually, we want to avoid this, but here it's the best way to get the required specificity without getting mad or resorting to !important-flagging.